test : 회원 정보 수정 api test를 다른 member test와 분리
- 이름 + 학번 수정을 프로필 수정으로 변경한 사항 적용

issue : #132
